+++ site/src/other.rs
@@ -4,7 +4,7 @@ use uiua::{CONSTANTS, ConstClass, Primitive, SysOp};
 use uiua_editor::{Editor, lang};
 
 use crate::{
-    Const, Hd, Prim, Prims,
+    Const, Hd, Hd3, Prim, Prims,
     markdown::{Markdown, markdown_view},
     primitive::doc_line_fragments_to_view,
 };
@@ -683,6 +683,11 @@ pub fn Subscripts() -> impl IntoView {
             </tr>
             { numeric }
         </table>
+        <Hd3 id="implicit">"Implicit Subscripts"</Hd3>
+        <p>"Adding a numeric subscript to a named function will pass that number as the first argument to the function."</p>
+        <Editor example="F ← ×10+\nF₂ 5"/>
+        <p>"You can define special cases as functions after the original"</p>
+        <Editor example="F  ← ×10+\nF₀ ← ×100\nF₁ ← $\"It's _\"\nF₀ 5\nF₁ 5\nF₂ 5\nF₃ 5\n"/>
 
         <Hd id="sided">"Sided Subscripts"</Hd>
         <p>"Sided subscripts augment a modifier or function in a way that can be thought of as having a \"side\"."</p>