+++ changelog.md
@@ -5,6 +5,9 @@ Uiua is not yet stable.
 ## 0.19.0 - 2026-??-??
 This version is not yet released. If you are reading this on the website, then these changes are live here.
 ### Language
+- Add custom subscripts
+  - [Implicit Subscripts](<https://www.uiua.org/docs/subscripts#implicit>) are stable
+  - [Custom subscript functions](<https://www.uiua.org/docs/experimental#custom-subscript-functions>) are experimental
 - Add [`oklch`](https://uiua.org/docs/oklch) function
 - Add [`un °`](https://uiua.org/docs/un)[`reshape ↯`](https://uiua.org/docs/reshape) to output the deshaped array alongside its shape